Specify enum names for improvement status values
product-improvement is the only use of `asPipelineItem` that requires enumNames to be overwritten.

Note - as a related fix we should update jellyfish-ui-component's getViewSlices helper method to make use of the `enumNames` (if defined) to return the appropriate value name (label) as well. That way we can make use of it when displaying the Status dropdown component in the View header.

Change-type: minor
Signed-off-by: Graham McCulloch <graham@balena.io>