Move jellyfish threads view to the product-os plugin
The jellyfish threads view is distinctly a product-os specific contract
as it displays support threads *about* product-os.
It was moved into the product-os plugin as part of this PR https://github.com/product-os/jellyfish-plugin-product-os/pull/472
Long term, there is still an open question about wether or not
product-os features are part of the core Jellyfish offering, or just
a use case, but I will defer that discussion to a later time.

Change-type: major
Signed-off-by: Lucian Buzzo <lucian.buzzo@gmail.com>