Decay pattern weight over time
This change uses the same decay algorithm as Hackernews:
```
Score = (P) / (T+2)^G

where,
P = points of an item (number of times contract is linked to)
T = time since creation (in hours)
G = Gravity, defaults to 1.8 in news.arc
```

This is re-calculated whenever the pattern is updated, essentially
a "poor mans cron". Once scheduled actions are implemented, we should be
able to recalculate the weight at explicit intervals.

Change-type: minor
Signed-off-by: Lucian Buzzo <lucian.buzzo@gmail.com>